+++ backend/controllers/game_controller.py
@@ -7,28 +7,19 @@ from fastapi import APIRouter, HTTPException, Depends, status
 
 from ..dto.game_dto import (
     StartGameRequest, StartGameResponse, SubmitAnswerRequest, SubmitAnswerResponse,
-    GameStatusResponse, FinishGameResponse, RuleDetailsResponse
+    GameStatusResponse, FinishGameResponse
 )
 from ..services.interfaces import IGameService
 from ..dependencies import get_game_service
 
-# Create combined router for backward compatibility
-router = APIRouter()
-
-game_router = APIRouter(
+router = APIRouter(
     prefix="/game", 
     tags=["Game Management"],
     responses={404: {"description": "Game session not found"}}
 )
 
-rules_router = APIRouter(
-    prefix="/rules", 
-    tags=["Grammar Rules"],
-    responses={404: {"description": "Rule not found"}}
-)
 
-
-@game_router.get(
+@router.get(
     "/start", 
     response_model=StartGameResponse,
     summary="Start New Game",
@@ -91,7 +82,7 @@ async def start_game(
         )
 
 
-@game_router.post(
+@router.post(
     "/{game_id}/step/{step_id}/answer", 
     response_model=SubmitAnswerResponse,
     summary="Submit Answer",
@@ -155,7 +146,7 @@ async def submit_answer(
         )
 
 
-@game_router.get(
+@router.get(
     "/{game_id}/status", 
     response_model=GameStatusResponse,
     summary="Get Game Status",
@@ -209,7 +200,7 @@ async def get_game_status(
         )
 
 
-@game_router.post(
+@router.post(
     "/{game_id}/finish", 
     response_model=FinishGameResponse,
     summary="Finish Game",
@@ -269,71 +260,3 @@ async def finish_game(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             detail=f"Error finishing game: {str(e)}"
         )
-
-
-@rules_router.get(
-    "/{sutra}", 
-    response_model=RuleDetailsResponse,
-    summary="Get Rule Details",
-    description="Retrieve comprehensive information about a specific Panini grammar rule."
-)
-async def get_rule_details(
-    sutra: str,
-    game_service: IGameService = Depends(get_game_service)
-) -> RuleDetailsResponse:
-    """
-    Get detailed information about a specific Panini grammar rule (sutra).
-    
-    Provides comprehensive reference information for Sanskrit grammar rules,
-    including descriptions, examples, and related rules for study and verification.
-    
-    **Path Parameters:**
-    - **sutra**: Panini rule number (e.g., "3.1.68", "1.4.14")
-    
-    **Returns:**
-    - **sutra**: The rule number as requested
-    - **description**: Sanskrit and English explanation of the rule
-    - **example**: Sample transformations demonstrating the rule
-    - **category**: Grammatical category (Lakara, Vibhakti, Sandhi, etc.)
-    - **next**: Related rule numbers for further study
-    
-    **Rule Categories:**
-    - **Lakara**: Tense and mood formations
-    - **Vibhakti**: Case endings and declensions
-    - **Sandhi**: Sound combination rules
-    - **Kridanta**: Verbal derivatives
-    - **Taddhita**: Nominal derivatives
-    
-    **Example:**
-    ```
-    GET /rules/3.1.68
-    ```
-    
-    **Response:**
-    ```json
-    {
-      "sutra": "3.1.68",
-      "description": "लट् लकारः — Present tense verbal endings",
-      "example": "गम् → गच्छति (he/she goes)",
-      "category": "Lakara",
-      "next": ["3.4.78", "3.1.69"]
-    }
-    ```
-    """
-    try:
-        return await game_service.get_rule_details(sutra)
-    except ValueError as e:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=str(e)
-        )
-    except Exception as e:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Error getting rule details: {str(e)}"
-        )
-
-
-# Include sub-routers in main router
-router.include_router(game_router)
-router.include_router(rules_router)
